refactored service and mapper files. verified API endpoints

+++ src/main/java/com/alex/barselplus_backend/model/Patient.java
@@ -31,7 +31,7 @@ public class Patient {
     private LocalDate dateOfBirth;
 
     @Column(name = "national_ID", unique = true, nullable = false)
-    private Integer nationalID;
+    private Long nationalID;
 
     @Column(name = "address", columnDefinition = "TEXT")
     private String address;
